refactor/docs(logger): Create technical document and add explanation for logger. Also create config file for logger and use it instead of a hardcoded object.
Creates an internal technical document in the GDD folder.
Creates a Logger.md file to explain how to use the logger service.
Creates a logger-config.json file so that each developer can have multiple logger configurations locally.
The logger-config will be added to the .gitignore in the next commit, but this commit will make sure everyone has it in the correct path.

BREAKING CHANGE: Now there has to be a logger-config.json in the src/configs folder for the logger to work.

refers to $1xa