feat(ítems & buffs): implement ítems levels depending on number of explored romos. Also implement buff merging for two buffs of the same type.
Refactors how item data is stored. Separates it into different files for clearer editing. Now for each item, an array of configs representing the different item levels is given.
Refactors item factory for the new way of storing ítems data. Also makes ítem keys store item level data to distinguish between the different levels.
Implements buff merging in the BuffManagerComponent, so that when to buffs of the same type are applied, they merge with whatever logic we want. Currently their duration is added and the highest value of both is kept.